requirements.txt
langchain
numpy
pandas
scikit-learn
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
PyMuPDF
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
fastapi
langserve
mysql-connector-python
SQLALchemy
validators
youtube_transcript_api
unstructured
numexpr
huggingface_hub
tiktoken
cassio 
astrapy
datasets
pyPDF2
pinecone-client
pinecone-text
pinecone-notebooks
neo4j
vllm
fastapi
uvicorn
pydantic
